airflow-requirements.txt
typing_extensions>=4.10.0
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.0
apache-airflow-providers-http==4.6.0
psycopg2-binary==2.9.9
Flask-Session==0.4.0
Django==4.2
djangorestframework
python-dateutil
